✨ 답변 메타데이터 추가
